WIP: Setup API integration for video invite dialog
